cluster: Unify map_errc
