+++ src/v/cluster/errc.cc
@@ -1,0 +1,52 @@
+// Copyright 2021 Vectorized, Inc.
+//
+// Use of this software is governed by the Business Source License
+// included in the file licenses/BSL.md
+//
+// As of the Change Date specified in that file, in accordance with
+// the Business Source License, use of this software will be governed
+// by the Apache License, Version 2.0
+
+#include "cluster/errc.h"
+
+#include "raft/errc.h"
+#include "rpc/errc.h"
+
+namespace cluster {
+
+cluster::errc map_errc(std::error_code ec) {
+    if (ec == errc::success) {
+        return errc::success;
+    }
+
+    // raft errors
+    if (ec.category() == raft::error_category()) {
+        switch (static_cast<raft::errc>(ec.value())) {
+        case raft::errc::timeout:
+            return errc::timeout;
+        case raft::errc::not_leader:
+            return errc::not_leader_controller;
+        default:
+            return errc::replication_error;
+        }
+    }
+
+    // from rpc errors
+    if (ec.category() == rpc::error_category()) {
+        switch (static_cast<rpc::errc>(ec.value())) {
+        case rpc::errc::client_request_timeout:
+            return errc::timeout;
+        default:
+            return errc::replication_error;
+        }
+    }
+
+    // cluster errors, just forward
+    if (ec.category() == cluster::error_category()) {
+        return static_cast<errc>(ec.value());
+    }
+
+    return errc::replication_error;
+}
+
+} // namespace cluster
